requirements.txt
Package                  Version
------------------------ -----------
aiohttp                  3.8.1
aiosignal                1.2.0
alpaca-trade-api         2.1.0
antlr4-python3-runtime   4.9.3
appnope                  0.1.2
asttokens                2.0.5
async-timeout            4.0.2
attrs                    21.4.0
backcall                 0.2.0
beautifulsoup4           4.11.1
cachetools               5.1.0
certifi                  2021.10.8
charset-normalizer       2.0.12
cronitor                 4.4.4
cycler                   0.11.0
debugpy                  1.5.1
decorator                5.1.1
deprecation              2.1.0
dropbox                  11.31.0
entrypoints              0.4
executing                0.8.3
fonttools                4.33.3
frozenlist               1.3.0
google-api-core          2.8.0
google-api-python-client 2.48.0
google-auth              2.6.6
google-auth-httplib2     0.1.0
google-auth-oauthlib     0.5.1
googleapis-common-protos 1.56.1
httplib2                 0.20.4
humanize                 4.1.0
hydra-core               1.2.0
idna                     3.3
ipykernel                6.9.1
ipython                  8.2.0
jedi                     0.18.1
jupyter-client           7.2.2
jupyter-core             4.9.2
kiwisolver               1.4.2
lxml                     4.8.0
matplotlib               3.5.2
matplotlib-inline        0.1.2
msgpack                  1.0.3
multidict                6.0.2
mysqlclient              2.1.0
nest-asyncio             1.5.5
numexpr                  2.8.1
numpy                    1.22.3
oauthlib                 3.2.0
omegaconf                2.2.2
packaging                21.3
pandas                   1.4.2
pandas-datareader        0.10.0
parso                    0.8.3
patsy                    0.5.2
pexpect                  4.8.0
pickleshare              0.7.5
Pillow                   9.1.1
pip                      21.2.4
ply                      3.11
prompt-toolkit           3.0.20
protobuf                 3.20.1
psycopg2                 2.8.6
ptyprocess               0.7.0
pure-eval                0.2.2
pyasn1                   0.4.8
pyasn1-modules           0.2.8
Pygments                 2.11.2
pygsheets                2.0.5
pyparsing                3.0.9
python-dateutil          2.8.2
pytz                     2022.1
PyYAML                   6.0
pyzmq                    22.3.0
requests                 2.27.1
requests-oauthlib        1.3.1
rsa                      4.8
scipy                    1.8.1
seaborn                  0.11.2
setuptools               61.2.0
six                      1.16.0
soupsieve                2.3.2.post1
stack-data               0.2.0
statsmodels              0.13.2
stone                    3.3.1
ta                       0.10.1
tables                   3.7.0
tornado                  6.1
traitlets                5.1.1
uritemplate              4.1.1
urllib3                  1.26.9
wcwidth                  0.2.5
websocket-client         1.3.2
websockets               10.3
wheel                    0.37.1
yarl                     1.7.2
